+++ review/core/lib/util/review_helper.dart
@@ -1,0 +1,4 @@
+
+abstract class ReviewHelper {
+  Future<void> requestReview();
+}